# api/openapi.yaml
openapi: 3.1.0
info:
  title: llm-proxy web API
  version: 0.1.0
  description: |
    The browser-facing API of llm-proxy: the user cabinet and the admin panel.
    This is NOT the proxied LLM API (`/v1/...`), which the embedded upstream serves
    and which authenticates with API tokens, not sessions.

    Conventions:
    - Authentication is a session cookie set by `POST /api/auth/login` or the OIDC
      callback. It is `HttpOnly`, so no endpoint ever returns a session id.
    - A restricted session (temporary password not yet changed) may call only
      `GET /api/me`, `POST /api/auth/password` and `POST /api/auth/logout`; every
      other authenticated endpoint answers 403 with code `password_change_required`.
    - An authenticated endpoint called without a valid session answers 401 with code
      `unauthenticated`. The only other 401 code is `invalid_credentials`, and it is
      returned only by `POST /api/auth/login` and `POST /api/auth/password`, where it
      rejects what was typed and leaves the session alone: a client must not treat it
      as a signed-out session.
    - Every error the web API writes, including those from its middleware (session,
      restriction, content type), is a JSON `Error` - never plain text.
    - Codes any endpoint may answer besides those it declares: `unsupported_media_type`
      (415, a mutating request without `Content-Type: application/json`),
...
      expects, with `field` when one field is at fault).
...
servers:
  - url: /
...
paths:
...
  /api/auth/oidc/start:
    get:
      tags: [auth]
      operationId: startOIDC
      summary: Redirects the browser to the identity provider.
      security: []
      responses:
        "302": { description: Redirect to the identity provider. }
        "404":
          description: "`oidc_disabled`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/oidc/callback:
    get:
      tags: [auth]
      operationId: oidcCallback
      summary: >
        Identity-provider redirect target. On success sets the session cookie and
        redirects to `/`; on refusal redirects to `/login?error=<code>` with one of
        `oidc_forbidden`, `oidc_failed`.
      security: []
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        "302": { description: Redirect back into the application. }
...
  /api/auth/password:
    post:
      tags: [auth]
      operationId: changePassword
      summary: >
        Change the caller's password. `currentPassword` is required unless the
        session is restricted (the temporary password was proved at sign-in).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordChangeRequest" }
      responses:
        "204": { description: Changed. A restricted session becomes a full one. }
        "400":
          description: "`weak_password` or `empty_password`, with `field`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: "`invalid_credentials` — the current password is wrong or the temporary one expired."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/admin/users:
    get:
      tags: [admin-users]
      operationId: listUsers
      responses:
        "200":
          description: Every account, humans and service accounts.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUser" }
    post:
      tags: [admin-users]
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: >
            Created. For a human with local sign-in, `temporaryPassword` is set and
            shown only here. A service account and an OIDC-only invitation get none.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedUser" }
        "409":
          description: "`email_taken` — compared case-insensitively."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: >
            `invalid_input` with `field` naming the offending field (kind, email, role,
            signIn, displayName), or `invalid_rule` with `field` set to the first bad rule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/admin/users/{userId}/invitation:
    post:
      tags: [admin-users]
      operationId: renewUserInvitation
      summary: >
        Issue a fresh identity-provider invitation for a human who has not linked an
        identity yet (the old one lapsed or was consumed without a link). Replaces any
        previous invitation for that account.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "204": { description: Invited. The first IdP sign-in with the account's address claims it. }
        "409":
          description: >
            `already_linked` — the account already has an identity-provider link;
            `not_invitable` — a service account, an account without an email address, or
            identity-provider sign-in is not configured.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: llmproxy_session
  parameters:
    UserId: { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine code. The client translates this, not `message`.
          examples: [invalid_credentials]
        message: { type: string, description: English diagnostic for logs. Not for display. }
        field: { type: string, description: "The request field the error is about, when there is one." }
...
    PasswordChangeRequest:
      type: object
      required: [newPassword]
      properties:
        currentPassword: { type: string, description: Required on a full session. }
        newPassword: { type: string, minLength: 12 }
    Kind: { type: string, enum: [human, service] }
    Role: { type: string, enum: [user, admin] }
    Status: { type: string, enum: [active, blocked] }
    PolicySource:
      type: string
      enum: [local, idp]
      description: "`idp` means the policy is recomputed from group claims at every sign-in and is read-only here."
    Policy:
      type: array
      description: >
        Allow-list rules `<provider>:<model-glob>`. The glob dialect is exactly `*`
        (any run of characters, including `/` and `:`) and `?` (one character);
        everything else is literal. A wildcard covers models published later.
      items: { type: string, examples: ["chatgpt:*", "claude:claude-sonnet-5"] }
...
    AdminUser:
      type: object
      required: [id, kind, displayName, role, status, policy, policySource, mustChangePassword, signIn, createdAt]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/Kind" }
        email: { type: [string, "null"] }
        displayName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        status: { $ref: "#/components/schemas/Status" }
        policy: { $ref: "#/components/schemas/Policy" }
        policySource: { $ref: "#/components/schemas/PolicySource" }
        mustChangePassword: { type: boolean }
        signIn:
          type: array
          description: >
            Sign-in methods that work now: `password` when a usable (unexpired) password is
            set, `oidc` when an identity-provider identity is linked. Empty for a service
            account and for a person whose only way in is a pending invitation.
          items: { type: string, enum: [password, oidc] }
        invitationExpiresAt:
          type: [string, "null"]
          format: date-time
          description: >
            Set while an identity-provider invitation exists and no identity is linked yet;
            it may lie in the past (lapsed - renew it with POST .../invitation).
        lastSeenAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    CreateUserRequest:
      type: object
      required: [kind, displayName, policy]
      properties:
        kind: { $ref: "#/components/schemas/Kind" }
        email: { type: string, description: Required for a human. Absent for a service account. }
        displayName: { type: string, minLength: 1 }
        role: { $ref: "#/components/schemas/Role" }
        policy: { $ref: "#/components/schemas/Policy" }
        signIn:
          type: string
          enum: [password, oidc]
          description: >
            For a human only. `password` issues a temporary password; `oidc` records an
            invitation the first sign-in with that address through the identity provider
            will claim. Ignored for a service account.
    CreatedUser:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/AdminUser" }
        temporaryPassword: { $ref: "#/components/schemas/TemporaryPassword" }
    TemporaryPassword:
      type: object
      required: [password, expiresAt]
      properties:
        password: { type: string, description: Shown once. }
        expiresAt: { type: string, format: date-time }
...
security:
  - session: []